Add Input Class (#39)
This makes it easier to access the current value of the input
element because you can access the `Input.value` attribute
instead of creating a `Var` and assigning the value to it via
an event handler.

To accomplish this the `Events` object needed to be refactored
because only one callback could be registered to a particular
event type. The `EventHandler` class now stores one or more
callback functions and triggers each when the event occurs.