Track contexts in hooks as state (#787)
* fix use_context

components which are newly added to the layout after the initial
render do not have access to contexts. we solve this by tracking
life cycle hooks in stack and copying the context providers from
parent to child hooks.

we also change how contexts are implemented - instead of needing
a create_context function which returns a new Context class, we
just return a Context object that constructs a ContextProvider
component. the earlier implementation was too clever and did not
add anything for it.

* add test

* add changelog entry

* fix mypy

* fix tornado dev server

* remove unused func

* improve test

* remove unused method

* fix style issues

* get clever with functions

* fix tests

* update changelog

* get coverage